refactor: use ark pricing service for rates
