+++ src/app/lib/ArkPricing/contracts.ts
@@ -1,0 +1,15 @@
+export interface PriceResponse {
+  data: {
+    coin: string;
+    prices: Partial<
+      Record<
+        string,
+        {
+          price: number;
+          change24h: number | null;
+          timestamp: string;
+        }
+      >
+    >;
+  };
+}
