--- src/app/lib/Coingecko/contracts.ts
@@ -1,7 +1,0 @@
-import { Coin } from "@/app/lib/Network";
-
-export interface MarketDataResponseItem {
-  name: Coin;
-  current_price: number;
-  id: string;
-}
